+++ Week06/정희선/readme.md
@@ -1,0 +1,7 @@
+# Algo-smash Week 06 스터디
+
+## 이번 주 스터디
+백트래킹 문제풀이
+
+
+발표 자료 - [https://velog.io/@lissani/Backtracking-%EA%B0%9C%EB%85%90]
