+++ Week07/정희선/study.md
@@ -1,0 +1,1 @@
+발표자료: https://www.notion.so/30f860002c5d80c488cbebc2fb7f2664?source=copy_link
