--- 백준/Gold/9663. N－Queen/README.md
@@ -1,30 +1,0 @@
-# [Gold IV] N-Queen - 9663 
-
-[문제 링크](https://www.acmicpc.net/problem/9663) 
-
-### 성능 요약
-
-메모리: 2020 KB, 시간: 5944 ms
-
-### 분류
-
-브루트포스 알고리즘, 백트래킹
-
-### 제출 일자
-
-2026년 2월 15일 04:24:02
-
-### 문제 설명
-
-<p>N-Queen 문제는 크기가 N × N인 체스판 위에 퀸 N개를 서로 공격할 수 없게 놓는 문제이다.</p>
-
-<p>N이 주어졌을 때, 퀸을 놓는 방법의 수를 구하는 프로그램을 작성하시오.</p>
-
-### 입력 
-
- <p>첫째 줄에 N이 주어진다. (1 ≤ N < 15)</p>
-
-### 출력 
-
- <p>첫째 줄에 퀸 N개를 서로 공격할 수 없게 놓는 경우의 수를 출력한다.</p>
-
