--- 백준/Silver/10828. 스택/README.md
@@ -1,38 +1,0 @@
-# [Silver IV] 스택 - 10828 
-
-[문제 링크](https://www.acmicpc.net/problem/10828) 
-
-### 성능 요약
-
-메모리: 2024 KB, 시간: 212 ms
-
-### 분류
-
-구현, 자료 구조, 스택
-
-### 제출 일자
-
-2026년 1월 25일 02:47:12
-
-### 문제 설명
-
-<p>정수를 저장하는 스택을 구현한 다음, 입력으로 주어지는 명령을 처리하는 프로그램을 작성하시오.</p>
-
-<p>명령은 총 다섯 가지이다.</p>
-
-<ul>
-	<li>push X: 정수 X를 스택에 넣는 연산이다.</li>
-	<li>pop: 스택에서 가장 위에 있는 정수를 빼고, 그 수를 출력한다. 만약 스택에 들어있는 정수가 없는 경우에는 -1을 출력한다.</li>
-	<li>size: 스택에 들어있는 정수의 개수를 출력한다.</li>
-	<li>empty: 스택이 비어있으면 1, 아니면 0을 출력한다.</li>
-	<li>top: 스택의 가장 위에 있는 정수를 출력한다. 만약 스택에 들어있는 정수가 없는 경우에는 -1을 출력한다.</li>
-</ul>
-
-### 입력 
-
- <p>첫째 줄에 주어지는 명령의 수 N (1 ≤ N ≤ 10,000)이 주어진다. 둘째 줄부터 N개의 줄에는 명령이 하나씩 주어진다. 주어지는 정수는 1보다 크거나 같고, 100,000보다 작거나 같다. 문제에 나와있지 않은 명령이 주어지는 경우는 없다.</p>
-
-### 출력 
-
- <p>출력해야하는 명령이 주어질 때마다, 한 줄에 하나씩 출력한다.</p>
-
